VBID-10667 Support Expirations in the .NET Client
Adds a method to the Advertisement entity which allows for sending
an expiration. Refactored the method for sending PoP GET requests
to do so with minimal duplication.

Updates the tests to handle this case.

Most of the noise in this patch is around updating the assembly version
in docs.
Changes of note are in Advertisment.cs and AdvertisementTest.cs